+++ src/Aggregator/Application/UseCase/Command/Clear.php
@@ -5,19 +5,15 @@ declare(strict_types=1);
 namespace App\Aggregator\Application\UseCase\Command;
 
 /**
- * Class SaveArticle.
+ * Class Clear.
  *
  * @author bernard-ng <bernard@devscast.tech>
  */
-final readonly class SaveArticle
+final readonly class Clear
 {
     public function __construct(
-        public string $title,
-        public string $link,
-        public string $categories,
-        public string $body,
         public string $source,
-        public int $timestamp
+        public ?string $category = null
     ) {
     }
 }