+++ src/Aggregator/Application/UseCase/Query/GetStatsQuery.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Aggregator\Application\UseCase\Query;
+
+/**
+ * Class GetStatsQuery.
+ *
+ * @author bernard-ng <bernard@devscast.tech>
+ */
+final class GetStatsQuery
+{
+}