+++ src/Aggregator/Domain/Entity/SupportedSources.php
@@ -12,7 +12,7 @@ namespace App\Aggregator\Domain\Entity;
 enum SupportedSources: string
 {
     case ACTUALITE = 'actualite.cd';
-    case POLITICO = 'politico.cd';
+    case BETO = 'beto.cd';
     case SEPT_SUR_SEPT = '7sur7.cd';
     case RADIO_OKAPI = 'radiookapi.net';
     case MEDIA_CONGO = 'mediacongo.net';