+++ src/Aggregator/Domain/Repository/ArticleRepository.php
@@ -24,7 +24,9 @@ interface ArticleRepository
 
     public function getByLink(string $link): ?Article;
 
-    public function export(?string $source, ?DateRange $date): array;
+    public function getByHash(string $hash): ?Article;
 
-    public function getLastCrawlDate(string $source, ?string $category): string;
+    public function export(?string $source, ?DateRange $date): \Generator;
+
+    public function clear(string $source, ?string $category): int;
 }