+++ src/Aggregator/Domain/Service/DateParser.php
@@ -58,10 +58,15 @@ final readonly class DateParser
             );
         }
 
+        if ($format === 'c') {
+            $date = str_replace('t', ' ', $date);
+            $format = 'Y-m-d H:i:s';
+        }
+
         $datetime = \DateTime::createFromFormat($format ?? self::DEFAULT_DATE_FORMAT, $date);
 
         return $datetime !== false ?
             $datetime->format('U') :
-            (new \DateTime())->format('U');
+            (new \DateTime('midnight'))->format('U');
     }
 }