+++ src/Aggregator/Presentation/Console/UpdateCommand.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace App\Aggregator\Presentation\Console;
 
-use App\Aggregator\Application\UseCase\Query\GetLastCrawlDateQuery;
+use App\Aggregator\Application\UseCase\Query\GetEarliestPublicationDateQuery;
+use App\Aggregator\Application\UseCase\Query\GetLatestPublicationDateQuery;
 use App\Aggregator\Domain\ValueObject\DateRange;
+use App\Aggregator\Domain\ValueObject\Direction;
 use App\Aggregator\Domain\ValueObject\FetchConfig;
 use App\Aggregator\Infrastructure\Crawler\Source\SourceFetcher;
 use App\SharedKernel\Application\Bus\QueryBus;
@@ -37,6 +39,7 @@ class UpdateCommand extends Command
     {
         $this->addArgument('source', InputArgument::REQUIRED, 'the website source to crawle');
         $this->addOption('category', null, InputOption::VALUE_OPTIONAL, 'the category to crawle');
+        $this->addOption('direction', null, InputOption::VALUE_OPTIONAL, 'the direction to crawle', 'forward', ['forward', 'backward']);
     }
 
     #[\Override]
@@ -54,12 +57,21 @@ class UpdateCommand extends Command
         /** @var string|null $category */
         $category = $input->getOption('category');
 
-        /** @var string $lastUpdate */
-        $lastUpdate = $this->queryBus->handle(new GetLastCrawlDateQuery($source, $category));
-        $today = (new \DateTimeImmutable())->format('Y-m-d H:i:s');
-        $range = DateRange::from(sprintf('%s--%s', $lastUpdate, $today), 'Y-m-d H:i:s', '--');
+        /** @var string $direction */
+        $direction = $input->getOption('direction');
+        $direction = Direction::from($direction);
 
-        $this->io->title(sprintf('Updating database with article from %s to %s', $lastUpdate, $today));
+        /** @var \DateTimeImmutable $date */
+        $date = $this->queryBus->handle(match ($direction) {
+            Direction::FORWARD => new GetLatestPublicationDateQuery($source, $category),
+            Direction::BACKWARD => new GetEarliestPublicationDateQuery($source, $category),
+        });
+
+        $range = $direction === Direction::FORWARD ?
+            DateRange::forward($date) :
+            DateRange::backward($date);
+
+        $this->io->title(sprintf('[%s] Updating with range %s', $direction->value, $range->format()));
         $this->sourceFetcher->fetch(new FetchConfig($source, date: $range, category: $category));
         $this->io->success('website crawled successfully');